+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 - `sandstone/FixedPopupPanels` to support accessibility properly
 - `sandstone/TabLayout` performance when focusing items in the layout
+- `sandstone/Input` to update `invalidTooltip` to the latest design
 
 ## [1.0.0-beta.6] - 2020-06-08